Stand by

+++ Code/CCapSensors.h
@@ -1,11 +1,13 @@
 #ifndef _CCAPSENSORS_H
 #define _CCAPSENSORS_H
 
+#include "freertos.h"
+
 class CCapSensors {
 public:
 	void initCapacitiveSensor();
 	void closeCapacitiveSensor();
-	int readCapSensors();
+	uint8_t readCapSensors();
 	CCapSensors static* getInstance();
 private:
 	CCapSensors static* instance;
